coderabbit suggested refactor.  really great suggestion
